# swagger/src/main/java/kovteba/dto/Auth.java
package kovteba.dto;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;
import lombok.ToString;

@Data
@AllArgsConstructor
@NoArgsConstructor
@ToString
public class Auth {

    private String name;
    private String pass;

}