# swagger/src/main/java/kovteba/dto/UserNotFound.java
package kovteba.dto;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;
import lombok.ToString;

@Data
@AllArgsConstructor
@NoArgsConstructor
@ToString
public class UserNotFound {

    private String error;
}